+++ AAEmu.Game/Models/Game/Units/Unit.cs
@@ -1203,7 +1203,7 @@ public class Unit : BaseUnit, IUnit
                         var newEffect =
                             new Buff(this, this, new SkillCasterUnit(ObjId), buffTemplate, null, DateTime.UtcNow)
                             {
-                                AbLevel = itemLevels[setCount.Key]
+                                AbLevel = (ushort)itemLevels[setCount.Key]
                             };
                         Buffs.AddBuff(newEffect);
                         appliedBuffs.Add(bonus.BuffId);
@@ -1336,7 +1336,7 @@ public class Unit : BaseUnit, IUnit
             var newEffect =
                 new Buff(this, this, new SkillCasterUnit(), buffTemplate, null, DateTime.UtcNow)
                 {
-                    AbLevel = (uint)gradeBuffAbLevel
+                    AbLevel = (ushort)gradeBuffAbLevel
                 };
 
             Buffs.AddBuff(newEffect);
@@ -1381,7 +1381,7 @@ public class Unit : BaseUnit, IUnit
                     var newEffect =
                         new Buff(this, this, new SkillCasterUnit(), itemAddedBuff, null, DateTime.UtcNow)
                         {
-                            AbLevel = (uint)itemAdded.Template.Level
+                            AbLevel = (ushort)itemAdded.Template.Level
                         };
 
                     Buffs.AddBuff(newEffect);
@@ -1417,7 +1417,7 @@ public class Unit : BaseUnit, IUnit
                         var newEffect =
                             new Buff(this, this, new SkillCasterUnit(), itemAddedChargedBuff, null, DateTime.UtcNow)
                             {
-                                AbLevel = (uint)itemAdded.Template.Level
+                                AbLevel = (ushort)itemAdded.Template.Level
                             };
                         Buffs.AddBuff(newEffect);
                     }
@@ -1440,7 +1440,7 @@ public class Unit : BaseUnit, IUnit
                     var newEffect =
                         new Buff(this, this, new SkillCasterUnit(), buffTemplate, null, DateTime.UtcNow)
                         {
-                            AbLevel = (uint)item.Template.Level
+                            AbLevel = (ushort)item.Template.Level
                         };
 
                     Buffs.AddBuff(newEffect);
@@ -1474,7 +1474,7 @@ public class Unit : BaseUnit, IUnit
                         var newEffect =
                             new Buff(this, this, new SkillCasterUnit(), itemAddedChargedBuff, null, DateTime.UtcNow)
                             {
-                                AbLevel = (uint)item.Template.Level
+                                AbLevel = (ushort)item.Template.Level
                             };
                         Buffs.AddBuff(newEffect);
                     }
